add check for 'black' state
Add a check for the rgbw value of the strip. If both the primary and secondary values are set to black (0000), then don't bother turning on the strip and doing the animation loop.

This is mainly useful for a relay module, since the relay will click when toggling on and off. Leaving it off means there is no power going to the strip for no reason when its black/off anyway, and the relay isn't clicking for no reason either.