add option for lux sensor bh1750
Add a new usermod option and compile flag for lux sensor. If the flag is
set then a new option will appear in the usermod settings allowing you
to set a lux threshold.
If the current lux is above the threshold, the strip won't turn on when
triggered by the PIR sensors.

This requires the BH1750 usermod and deps, see that one for more details
on installation.